Fix ipxtester build failure.
This has been failing for a while due to the kernel in the base
Debian 10 image not matching the latest one in the repo, so the
kernel headers metapackage pulls in the wrong ones.

Could figure out a way to select the correct headers package, or
just install 11 which seems to work fine.